Develop

+++ shop-client/src/pages/CategoryDetails.tsx
@@ -1,18 +1,24 @@
-import { Paper, Typography } from '@mui/material';
+import { Paper, Typography, useMediaQuery, useTheme } from '@mui/material';
 import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useAppContext, useToastContext } from '../context';
 import { CategoryService } from '../services';
 import { Category } from '../types';
 import { ActionButtons } from '../components';
+import { useAppDispatch } from '../context/hooks';
+import { setToast } from '../context/ToastSlice';
+import { handleAction } from '../utils/actionHandler';
 
 const CategoryDetails = () => {
     const { id } = useParams();
     const navigate = useNavigate();
+    const dispatch = useAppDispatch();
     const { setLoading } = useAppContext();
-    const { setToast } = useToastContext();
     const [category, setCategory] = useState<Category | null>(null);
 
+    const theme = useTheme();
+    const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+    
     const getCategory = (categoryId: string) => {
         CategoryService.getCategory(categoryId).then((res) => {
             setCategory(res.data);
@@ -24,19 +30,15 @@ const CategoryDetails = () => {
     }, [id]);
 
     const handleDelete = () => {
-        setLoading(true);
-        id &&
-            CategoryService.deleteCategory(id)
-                .then(() => {
-                    navigate('/category');
-                    setToast({ severity: 'success', message: 'La catégorie a bien été supprimée' });
-                })
-                .catch(() => {
-                    setToast({ severity: 'error', message: 'Une erreur est survenue lors de la suppresion' });
-                })
-                .finally(() => {
-                    setLoading(false);
-                });
+        if (!id) return; 
+        handleAction(
+            CategoryService.deleteCategory(id),
+            'La catégorie a bien été supprimée',
+            '/',
+            dispatch,
+            navigate,
+            setLoading
+        );
     };
 
     const handleEdit = () => {
@@ -55,7 +57,7 @@ const CategoryDetails = () => {
         >
             <ActionButtons handleDelete={handleDelete} handleEdit={handleEdit} />
 
-            <Typography variant="h3" sx={{ textAlign: 'center' }}>
+            <Typography variant={isMobile ? "h5":"h3"} sx={{ textAlign: 'center' }}>
                 {category.name}
             </Typography>
         </Paper>
